Add Slider library and components, reorganize files
