+++ src/components/App/App.css
@@ -1,0 +1,21 @@
+.app {
+  /* display: flex;
+  flex-direction: column;
+  align-items: center; */
+  height: 98vh;
+  width: 95vw;
+}
+
+.all-content {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  height: 100%;
+
+}
+
+
+
+p {
+  margin: 0%;
+}
