+++ src/components/CitiesContainer/CitiesContainer.css
@@ -1,0 +1,7 @@
+.cities-container {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-gap: 1em;
+  width: 100%;
+  margin: 1%;
+}
