+++ src/components/CityCards/CityCards.css
@@ -1,0 +1,7 @@
+.card-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  border: 2px solid crimson;
+}
