--- src/components/CitySlider.js
@@ -1,13 +1,0 @@
-import React from "react";
-import '../styles/CitySlider.css';
-
-const CitySlider = () => {
-
-  return(
-    <article className='slider-container'>
-      <img src='https://upload.wikimedia.org/wikipedia/commons/thumb/e/e8/Denver_skyline.jpg/1200px-Denver_skyline.jpg' alt='Denver Colorado' />
-    </article>
-  )
-};
-
-export default CitySlider;
