+++ src/components/CitySlider/CitySlider.css
@@ -1,0 +1,15 @@
+.slider-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 25%;
+  width: 80%;
+  border: 2px solid crimson;
+}
+
+img {
+  display: block;
+  max-height: 100%;
+  width: 100%;
+  object-fit: cover;
+}
