+++ src/components/MainContent/MainContent.css
@@ -1,0 +1,7 @@
+.main-content {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+}
