+++ src/components/MainContent/MainContent.js
@@ -1,0 +1,20 @@
+// import React from "react";
+// import './MainContent.css';
+// import CitiesContainer from "../CitiesContainer/CitiesContainer";
+// import CitySlider from "../CitySlider/CitySlider";
+// import SearchBar from "../SearchBar/SearchBar";
+// // import { mockData } from "../../mockData";
+
+// const MainContent = () => {
+
+//   return(
+//     <section className='main-content'>
+//         <CitySlider />
+//       <SearchBar />
+//       <h2>All Cities</h2>
+//       <CitiesContainer />
+//     </section>
+//   )
+// };
+
+// export default MainContent;
