+++ src/components/Nav/Nav.css
@@ -1,0 +1,7 @@
+.navigation {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  height: 10%;
+  width: 100%;
+}
