+++ src/components/Nav/Nav.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../styles/Nav.css';
+import './Nav.css';
 
 const Nav = () => {
