+++ src/components/SearchBar/SearchBar.js
@@ -1,5 +1,5 @@
 import React from "react";
-import '../styles/SearchBar.css';
+import './SearchBar.css';
 
 const SearchBar = () => {
